+++ src/main/java/HollowKnight/model/game/scene/Scene.java
@@ -122,8 +122,8 @@ public class Scene {
 
     public boolean isEmpty(Position position) {
         for (Tile tile : tiles) {
-            int tileX = tile.getPosition().x();
-            int tileY = tile.getPosition().y();
+            int tileX = (int)tile.getPosition().x();
+            int tileY = (int)tile.getPosition().y();
 
             // Check if the position is within the bounds of the 8x8 tile
             if (position.x() >= tileX && position.x() < tileX + 8 &&
